chore: codify caching policy across routes
