+++ app/privacy/page.mdx
@@ -1,3 +1,8 @@
+import { REVALIDATION_WINDOWS } from "@/config/cache"
+
+export const dynamic = "force-static"
+export const revalidate = REVALIDATION_WINDOWS.MARKETING
+
 # Privacy Policy
 
 *Last Updated: [Current Date]*
